[core] Move ABI constants to `chains/abi`
